requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.0
cryptography==37.0.4
discord.py==1.7.3
distlib==0.3.5
filelock==3.7.1
frozenlist==1.3.0
h11==0.13.0
idna==3.3
multidict==6.0.2
outcome==1.2.0
platformdirs==2.5.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dotenv==0.20.0
requests==2.0.0
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.10
virtualenv==20.15.1
wsproto==1.1.0
yarl==1.7.2